requirements.txt
PyYAML
requests
urllib3
